feat: изменена схема бд

+++ src/Champ.API/Models/Doctor.cs
@@ -6,7 +6,7 @@ namespace Champ.API.Models;
 public class Doctor {
     [Key]
     [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-    public int DoctorId { get; set; } = 0;
+    public long DoctorId { get; set; } = 0;
 
     [MaxLength(255)]
     public string FirstName { get; set; } = string.Empty;
